+++ crates/server/src/domain/quests.rs
@@ -70,47 +70,6 @@ pub async fn start_quest(
     Ok(db.start_quest(quest_id, user_address).await?)
 }
 
-pub async fn reset_quest_instance(
-    db: Arc<impl QuestsDatabase>,
-    auth_user_address: &str,
-    quest_instance_id: &str,
-) -> Result<(), QuestError> {
-    match db.get_quest_instance(quest_instance_id).await {
-        Ok(instance) => match db.get_quest(&instance.quest_id).await {
-            Ok(quest) => {
-                if !auth_user_address.eq_ignore_ascii_case(&quest.creator_address) {
-                    return Err(QuestError::ResetQuestInstanceNotAllowed);
-                }
-
-                // remove events to reset quest instance state
-                db.remove_events(quest_instance_id).await.map_err(|err| {
-                    let err: QuestError = err.into();
-                    err
-                })?;
-
-                db.remove_instance_from_completed_instances(quest_instance_id)
-                    .await
-                    .map_err(|err| {
-                        let err: QuestError = err.into();
-                        err
-                    })?;
-
-                Ok(())
-            }
-            Err(err) => {
-                log::error!("Error getting quest: {:?}", err);
-                let err: QuestError = err.into();
-                Err(err)
-            }
-        },
-        Err(err) => {
-            log::error!("Error getting quest instance: {:?}", err);
-            let err: QuestError = err.into();
-            Err(err)
-        }
-    }
-}
-
 impl From<QuestStateCalculationError> for QuestError {
     fn from(value: QuestStateCalculationError) -> Self {
         match value {
